refactor: divide ui into components & removing unused code
