+++ components/header.tsx
@@ -1,0 +1,17 @@
+export function Header() {
+  return (
+    <div className="py-8">
+      <div className="max-w-7xl mx-auto px-6">
+        <div className="text-center">
+          <h1 className="text-3xl md:text-4xl font-bold text-foreground mb-3">
+            Select Your University
+          </h1>
+          <p className="text-base text-muted-foreground max-w-2xl mx-auto">
+            Track countdowns to college decision dates and never miss an
+            important announcement
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
